+++ tests/test_teams.py
@@ -3,6 +3,8 @@ Implements unit tests to teams router.
 """
 
 # Python imports.
+import os
+from dotenv import load_dotenv
 from fastapi.testclient import TestClient
 from sqlmodel import Session
 
@@ -12,7 +14,15 @@ from app.database.database import engine
 from app.models import TeamDB
 
 
+load_dotenv('.env')
+
+
 client = TestClient(app)
+
+
+def get_token():
+    response = client.post('/auth/login', json={"username": os.getenv('USER'), "password": os.getenv('PASSWORD')})
+    return response.json()['token']
 
 
 def delete_created_db_record(model, id):
@@ -25,17 +35,17 @@ def delete_created_db_record(model, id):
 
 def test_create_team():
     team_to_create = {"name": "Name", "country": "Country", "city": "City", "stadium": "Stadium", "color": "Color", "coach": "Coach"}
-    response = client.post("/teams/", json=team_to_create)
+    response = client.post("/teams/", json=team_to_create, headers={"Authorization": get_token()})
     json_response = response.json()
-    # delete_created_db_record(TeamDB, json_response.pop("id"))
-    json_response.pop("id")
+    id = json_response.pop("id")
     assert response.status_code == 201
     assert json_response == team_to_create
+    # delete_created_db_record(TeamDB, id)
 
 
 def test_bad_create_team():
     team_to_create = {"team_name": "Name", "COUNTRY": "Country", "cities": "City", "manager": "Coach"}
-    response = client.post("/teams/", json=team_to_create)
+    response = client.post("/teams/", json=team_to_create, headers={"Authorization": get_token()})
     assert response.status_code == 422
     assert response.json()['detail'][0]['msg'] == 'Field required'
 
@@ -68,7 +78,7 @@ def test_bad_get_team_by_id():
 def test_update_team():
     team_id = client.get('/teams/').json()[0]['id']
     update_data = {"city": "London", "color": "Red/White"}
-    response = client.patch(f'/teams/{team_id}', json=update_data)
+    response = client.patch(f'/teams/{team_id}', json=update_data, headers={"Authorization": get_token()})
     assert response.status_code == 200
     assert response.json()['city'] == "London"
     assert response.json()['color'] == 'Red/White'
@@ -77,13 +87,13 @@ def test_update_team():
 
 def test_delete_team():
     team_id = client.get('/teams/').json()[0]['id']
-    response = client.delete(f'/teams/{team_id}')
+    response = client.delete(f'/teams/{team_id}', headers={"Authorization": get_token()})
     assert response.status_code == 200
     assert response.json()['id'] == team_id
 
 
 def test_get_players_by_team_id():
     team_id = client.get('/teams/').json()[0]['id']
-    response = client.get(f'/teams/{team_id}/players')
+    response = client.get(f'/teams/{team_id}/players', headers={"Authorization": get_token()})
     assert response.status_code == 200
     assert isinstance(response.json(), list)
